refactor: extract FieldSuggestChips from CardFactoryModal.tsx (#8608) [part 4]
Continue the bounded extractions from CardFactoryModal.tsx (follow-ups to
#8834, #8838, #8861).

Move the FieldSuggestChips inline subcomponent (the row of "+ field"
chips that proposes new T1 columns from the JSON sample data) into its
own sibling file, mirroring the pattern used for cardFactoryPreviews and
cardFactoryAiTypes. Pure code movement: zero behavior change.

- New: web/src/components/dashboard/FieldSuggestChips.tsx (73 LOC)
- CardFactoryModal.tsx: 1448 -> 1383 lines
- Drop now-unused detectFieldFormat and useAIMode imports from CFM

Verification:
- npx tsc --noEmit: clean (zero errors)
- eslint FieldSuggestChips.tsx: clean
- eslint CardFactoryModal.tsx: same 1 pre-existing error + 11 pre-existing
  warnings as on main (no new lint regressions)
- vitest CardFactoryModal-related tests: 28 passed (4 files)

Refs #8608
